+++ src/main/java/com/greeenai/greeenai/domain/auth/controller/AuthController.java
@@ -2,7 +2,7 @@ package com.greeenai.greeenai.domain.auth.controller;
 
 import com.greeenai.greeenai.domain.auth.dto.LoginResponse;
 import com.greeenai.greeenai.domain.auth.service.AuthService;
-import com.greeenai.greeenai.domain.member.dto.request.LoginRequest;
+import com.greeenai.greeenai.domain.auth.dto.request.LoginRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
