+++ src/main/java/com/greeenai/greeenai/domain/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.greeenai.greeenai.domain.member.dto.request;
+package com.greeenai.greeenai.domain.auth.dto.request;
 
 import com.greeenai.greeenai.domain.member.domain.OauthProvider;
 import jakarta.validation.constraints.NotBlank;
