+++ src/test/java/com/greeenai/greeenai/domain/auth/service/AuthServiceTest.java
@@ -10,7 +10,7 @@ import com.greeenai.greeenai.domain.auth.domain.RefreshToken;
 import com.greeenai.greeenai.domain.auth.dto.LoginResponse;
 import com.greeenai.greeenai.domain.auth.repository.RefreshTokenRepository;
 import com.greeenai.greeenai.domain.member.domain.Member;
-import com.greeenai.greeenai.domain.member.dto.request.LoginRequest;
+import com.greeenai.greeenai.domain.auth.dto.request.LoginRequest;
 import com.greeenai.greeenai.domain.member.repository.MemberRepository;
 import com.greeenai.greeenai.global.property.JwtProperties;
 import com.greeenai.greeenai.global.security.JwtService;
